requirements.txt
python-dotenv
inflection